% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_mibcCPI.R
\name{classify_mibcCPI}
\alias{classify_mibcCPI}
\title{Single-sample classifier of CPI-response related classes}
\usage{
classify_mibcCPI(gexp, gene_id = c("gene_symbol", "ensembl_id", "entrezid")[1])
}
\arguments{
\item{gexp}{gene expression matrix, normalized by log2(FPKM)}

\item{gene_id}{type of gene identifier. One of: 'gene_symbol',
'ensembl_id', 'entrezid'}
}
\value{
A data.frame with classification predictions and probabilities.
}
\description{
Classifies RNA-seq from bladder cancer samples normalized with log2(FPKM) into one of 5 classes
proposed by Robertson et al (2021, doi: tba), related to response to
neoadjuvant treatment with checkpoint inhibitors (CPI).
}
\examples{
data(tcga_blca_ex)
classify_mibcCPI(tcga_blca_ex)
}
