% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_mibcCPI.R
\docType{data}
\name{tcga_blca_ex}
\alias{tcga_blca_ex}
\title{A 100-gene, 81 sample subset of the TCGA-BLCA dataset}
\format{
A matrix with 100 rows (genes) and 81 columns (samples)
}
\usage{
tcga_blca_ex
}
\description{
A 100-gene, 81 sample subset of the TCGA-BLCA dataset
}
\references{
Robertson et al., Cell, 2017 doi: 10.1016/j.cell.2017.09.007
}
\keyword{datasets}
