% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_mibcCPI.R
\name{classification_features}
\alias{classification_features}
\title{Get mibcCPI classification features}
\usage{
classification_features(
  gene_id = c("hgnc_symbol", "ensembl_gene_id", "entrezid")[1]
)
}
\arguments{
\item{gene_id}{type of gene identifier. One of: 'hgnc_symbol',
'ensembl_gene_id', 'entrezid'. Default: 'hgnc_symbol}
}
\description{
Returns a vector of 100 gene symbols with the classification features
}
\examples{
classification_features()
}
